Lots changes, I've been really bad about commiting

+++ c_ext/include/plane.h
@@ -9,14 +9,14 @@
  */
 typedef struct {
 	Vector cross_product;
-	double d;
+	long double d;
 	Vector center;
 } Axis;
 
 
 typedef struct {
 	Vector cross_product;
-	double d;
+	long double d;
 	Vector center;
 	Axis x_axis;
 	Axis y_axis;
@@ -25,7 +25,7 @@ typedef struct {
 /**
  * @method constructs plane from a cross product vector and a 'd' value
  */
-Plane p_new(Vector, double);
+Plane p_new(Vector, long double);
 
 /**
  * @method constructs plane from 3 points in space
@@ -39,13 +39,13 @@ Plane p_from_cp_cent(Vector, Vector);
 
 Vector p_closest(Plane, Vector);
 
-double p_dist(Plane, Vector);
+long double p_dist(Plane, Vector);
 
 Vector p_project(Plane, Vector);
 
 Vector p_unproject(Plane, Vector);
 
-double p_evaluate(Plane, Vector);
+long double p_evaluate(Plane, Vector);
 
 void _p_string(char*, size_t, Plane);
